# visol/design/prototype/API.yaml
openapi: 3.0.0
info:
  title: ViSOL
  description: >-
    Visual Interactive Optimization Tool.
  contact: { }
  version: 1.0.0
servers:
  - url: 'https://visol1.paas.hosted-by-previder.com/api/v1'
    variables: { }
paths:
...
  /employees:
    post:
      tags:
        - Employees
      summary: Create an Employee
      operationId: createEmployee
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
      responses:
        '201':
          description: ''
          headers:
            Location:
              schema:
                type: string
                pattern: '^\/employees\/[A-Za-z0-9._+%-]+@[A-Za-z0-9.-]+[.][A-Za-z]+$'
                maxLength: 75
                example: /employees/jdoe@tormails.com
              description: The URL of the newly created Employee.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags:
        - Employees
      summary: Retrieve all Employees
      operationId: retrieveAllEmployees
      parameters: [ ]
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /employees/{employee_email}:
    parameters:
      - name: employee_email
        in: path
        description: >-
          The email address of the employee.
        required: true
        schema:
          $ref: '#/components/schemas/Email'
    get:
      tags:
        - Employees
      summary: Retrieve an Employee
      operationId: retrieveEmployee
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags:
        - Employees
      summary: Replace an Employee
      operationId: replaceEmployee
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags:
        - Employees
      summary: Delete an Employee
      operationId: deleteEmployee
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /ports:
    post:
      tags:
        - Ports
      summary: Create a Port
      operationId: createPort
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Port'
        required: true
      responses:
        '201':
          description: ''
          headers:
            Location:
              schema:
                type: string
                pattern: '^\/ports\/[0-9]{1,10}$'
                maxLength: 17
                example: /ports/1
              description: The URL of the newly created Port.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Port'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags:
        - Ports
      summary: Retrieve all Ports
      operationId: retrieveAllPorts
      parameters: [ ]
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortMap'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/ports/import':
    post:
      tags:
        - Ports
      summary: Import a Port
      operationId: importPort
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortDump'
        required: true
      responses:
        '201':
          description: ''
          headers:
            Location:
              schema:
                type: string
                pattern: '^\/ports\/[0-9]{1,10}$'
                maxLength: 17
                example: /ports/1
              description: The URL of the newly imported Port.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Port'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/ports/{port_id}':
    parameters:
      - name: port_id
        in: path
        description: ID of the Port.
        required: true
        schema:
          $ref: '#/components/schemas/PortID'
    get:
      tags:
        - Ports
      summary: Retrieve a Port
      operationId: retrievePort
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Port'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags:
        - Ports
      summary: Replace a Port
      operationId: replacePort
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Port'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Port'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags:
        - Ports
      summary: Delete a Port
      operationId: deletePort
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/ports/{port_id}/export':
    parameters:
      - name: port_id
        in: path
        description: ID of the Port.
        required: true
        schema:
          $ref: '#/components/schemas/PortID'
    get:
      tags:
        - Ports
      summary: Export a Port
      operationId: exportPort
      parameters:
        - name: from
          in: query
          description: Time to start exporting from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end exporting at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortDump'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/ports/{port_id}/terminals':
    parameters:
      - name: port_id
        in: path
        description: ID of the Port.
        required: true
        schema:
          $ref: '#/components/schemas/PortID'
    get:
      tags:
        - Ports
      summary: Retrieve a Port's Terminals
      operationId: retrievePortTerminals
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalMap'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/ports/{port_id}/schedules':
    parameters:
      - name: port_id
        in: path
        description: ID of the Port.
        required: true
        schema:
          $ref: '#/components/schemas/PortID'
    get:
      tags:
        - Ports
      summary: Retrieve a Port's Schedules
      operationId: retrievePortSchedules
      parameters:
        - name: from
          in: query
          description: Time to start looking from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end looking at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortSchedules'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/ports/{port_id}/performance':
    parameters:
      - name: port_id
        in: path
        description: ID of the Port.
        required: true
        schema:
          $ref: '#/components/schemas/PortID'
    get:
      tags:
        - Ports
      summary: Calculate a Port's Performance
      operationId: calculatePortPerformance
      parameters:
        - name: from
          in: query
          description: Time to start calculating from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end calculating at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Performance'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
security:
  - bearerAuth: [ ]
components:
  schemas:
...
    Port:
      title: Port
      description: |-
        A Port object has the following attributes:
        
        
        + name: `Han Shipping` (string) - The name of the Port.
      type: object
      properties:
        name:
          type: string
          pattern: '^(?:.|\s)+$'
          maxLength: 64
      additionalProperties: false
      required:
        - name
      example:
        name: Han Shipping
    PortMap:
      title: Port Map
      description: A map of Port objects keyed by Port ID.
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Port'
      example:
        '1':
          name: Han Shipping
        '2':
          name: Thai Transport
    PortDump:
      title: Port Dump
      description: >-
        A Port Dump contains all information about a Port (possibly restricted to a specific time range). It has the following attributes:
        
        + version: `1` (integer) - A single digit version of the Port Dump format.
        
        + name: `Han Shipping Scenario 3` (string) - The name of the Port Dump.
        
        + terminals: `{"terminal3":{"name":"North","port":"port2"}}` (json) - All Terminals of the Port.
        
        + berths: `{"berth5":{"open":"06:15:00","close":"19:50:00","depth":4,"width":5,"length":18,"terminal":"terminal3","unload_speed":16}}` (json) - All Berths that fall under a Terminal of the Port.
        
        + vessels: `{"vessel9":{"name":"Llandaff","depth":3,"width":5,"length":2,"arrival":"2022-06-23T12:30:00Z","deadline":"2022-06-25T23:00:00Z","containers":12,"destination":"terminal3","cost_per_hour":14}}` (json) - All Vessels scheduled for a Terminal of the Port (within the provided time range).
        
        + schedules: `[{"berth":"berth5","start":"2022-06-23T13:30:00Z","manual":true,"vessel":"vessel10","expected_end":"2022-06-23T14:15:00Z"}]` (json) - All Schedules of Berths that fall under a Terminal of the Port (within the provided time range).
      type: object
      properties:
        version:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
          example: 1
        name:
          type: string
          pattern: '^(?:.|\s)+$'
          maxLength: 64
          example: Han Shipping Scenario 3
        terminals:
          $ref: '#/components/schemas/TerminalMap'
        berths:
          $ref: '#/components/schemas/BerthMap'
        vessels:
          $ref: '#/components/schemas/VesselMap'
        schedules:
          $ref: '#/components/schemas/BerthSchedules'
      additionalProperties: false
      required:
        - version
        - name
        - terminals
        - berths
        - vessels
        - schedules
      example:
        version: 1
        name: New York Imported
        terminals:
          terminal7:
            name: North
            port: port5
          terminal8:
            name: South
            port: port5
        berths:
          berth13:
            open: '06:15:00'
            close: '19:50:00'
            depth: 4
            width: 5
            length: 18
            terminal: terminal7
            unload_speed: 16
          berth14:
            open: '09:50:00'
            close: '21:45:00'
            depth: 5
            width: 5
            length: 14
            terminal: terminal7
            unload_speed: 14
          berth15:
            open: '10:15:00'
            close: '22:15:00'
            depth: 7
            width: 5
            length: 15
            terminal: terminal8
            unload_speed: 12
          berth16:
            open: '07:45:00'
            close: '23:00:00'
            depth: 4
            width: 4
            length: 8
            terminal: terminal8
            unload_speed: 18
        vessels:
          vessel42:
            name: Llandaff
            depth: 3
            width: 5
            length: 2
            arrival: '2022-06-23T12:30:00Z'
            deadline: '2022-06-25T23:00:00Z'
            containers: 12
            destination: terminal7
            cost_per_hour: 14
          vessel43:
            name: Tang
            depth: 2
            width: 2
            length: 1
            arrival: '2022-06-23T17:05:06Z'
            deadline: '2022-06-24T23:15:00Z'
            containers: 12
            destination: terminal7
            cost_per_hour: 412
          vessel44:
            name: Jahde
            depth: 25
            width: 25
            length: 25
            arrival: '2022-06-23T12:05:06Z'
            deadline: '2022-06-23T23:05:06Z'
            containers: 24
            destination: terminal7
            cost_per_hour: 12
          vessel45:
            name: Landudno
            depth: 26
            width: 25
            length: 36
            arrival: '2022-06-23T13:05:06Z'
            deadline: '2022-06-23T23:05:06Z'
            containers: 13
            destination: terminal7
            cost_per_hour: 12
          vessel46:
            name: Picton
            depth: 26
            width: 25
            length: 38
            arrival: '2022-06-24T12:05:06Z'
            deadline: '2022-06-24T23:05:06Z'
            containers: 13
            destination: terminal8
            cost_per_hour: 3.2
          vessel47:
            name: Etna
            depth: 36
            width: 24
            length: 24
            arrival: '2022-06-24T13:05:06Z'
            deadline: '2022-06-24T23:05:06Z'
            containers: 45
            destination: terminal8
            cost_per_hour: 0.5
          vessel48:
            name: Wem
            depth: 64
            width: 24
            length: 14
            arrival: '2022-06-23T14:05:06Z'
            deadline: '2022-06-23T23:05:06Z'
            containers: 234
            destination: terminal8
            cost_per_hour: 0.01
          vessel49:
            name: The Cyane
            depth: 53
            width: 12
            length: 17
            arrival: '2022-06-23T16:05:06Z'
            deadline: '2022-06-23T23:05:06Z'
            containers: 5
            destination: terminal8
            cost_per_hour: 1.2
        schedules:
          - berth: berth13
            start: '2022-06-23T13:30:00Z'
            manual: true
            vessel: vessel43
            expected_end: '2022-06-23T14:15:00Z'
          - berth: berth13
            start: '2022-06-23T14:00:00Z'
            manual: true
            vessel: vessel42
            expected_end: '2022-06-23T14:45:00Z'
          - berth: berth14
            start: '2000-01-08T13:05:06Z'
            manual: true
            vessel: vessel45
            expected_end: '2000-01-08T14:00:48Z'
          - berth: berth14
            start: '2022-06-23T13:00:00Z'
            manual: false
            vessel: vessel44
            expected_end: '2022-06-23T14:42:51Z'
          - berth: berth15
            start: '2000-01-08T13:05:06Z'
            manual: true
            vessel: vessel47
            expected_end: '2000-01-08T16:50:06Z'
          - berth: berth15
            start: '2000-01-09T13:05:06Z'
            manual: false
            vessel: vessel46
            expected_end: '2000-01-09T14:10:06Z'
          - berth: berth16
            start: '2000-01-08T13:05:06Z'
            manual: false
            vessel: vessel49
            expected_end: '2000-01-08T13:21:46Z'
          - berth: berth16
            start: '2000-01-11T13:05:06Z'
            manual: false
            vessel: vessel48
            expected_end: '2000-01-12T02:05:06Z'
...
    Email:
      title: Email
      description: An email address for an Employee
      type: string
      format: email
      pattern: '^[A-Za-z0-9._+%-]+@[A-Za-z0-9.-]+[.][A-Za-z]+$'
      maxLength: 64
      example: jdoe@tormails.com
    EmployeeInput:
      title: Employee Input
      description: >-
        An Employee Input object has the following attributes:
        
        + email: `jdoe@tormails.com` (string) - The email address of the Employee.
        
        + password: `Password123` (string) - The password of the Employee.
        
        + role: `vessel planner` (enum) - The role of the Employee.
        
        + terminal: `1` (number|null) - The ID of the Employee's Terminal.
        
        + port: `null` (number|null) - The ID of the Employee's Port.
      type: object
      properties:
        email:
          $ref: '#/components/schemas/Email'
        password:
          type: string
          format: password
          pattern: '^(?:.|\s)*$'
          maxLength: 1073741824
          example: Password123
        role:
          $ref: '#/components/schemas/Role'
        terminal:
          allOf:
            - $ref: '#/components/schemas/TerminalID'
            - nullable: true
        port:
          allOf:
            - $ref: '#/components/schemas/PortID'
            - nullable: true
      additionalProperties: false
      required:
        - email
        - password
        - role
        - terminal
        - port
      example:
        email: jdoe@tormails.com
        password: Password123
        role: vessel planner
        terminal: 1
        port: null
    Employee:
      title: Employee
      description: >-
        An Employee object has the following attributes:
        
        + email: `jdoe@tormails.com` (string) - The email address of the Employee.
        
        + key
        
          + hash: `U+Q43l2oMyLFXJBgZOGo2AF5d3sy0lWQYjqX+bVEwm/KwBWSP83OVRHlbi3tDc6KdOXc6N7JzOkK9s50PYOjEg==` (byte) - The hash of the Employee's password.
        
          + salt: `9qtiDMW9KczgmJUhYaj/n+IrS0LoWBNZFWARUZV1RQk=` (byte) - The salt used to generate and validate the hash.
        
        + role: `vessel planner` (enum) - The role of the Employee.
        
        + terminal: `1` (number|null) - The ID of the Employee's Terminal.
        
        + port: `null` (number|null) - The ID of the Employee's Port.
      type: object
      properties:
        email:
          $ref: '#/components/schemas/Email'
        key:
          type: object
          properties:
            hash:
              type: string
              format: byte
              example: U+Q43l2oMyLFXJBgZOGo2AF5d3sy0lWQYjqX+bVEwm/KwBWSP83OVRHlbi3tDc6KdOXc6N7JzOkK9s50PYOjEg==
            salt:
              type: string
              format: byte
              example: 9qtiDMW9KczgmJUhYaj/n+IrS0LoWBNZFWARUZV1RQk=
          additionalProperties: false
          required:
            - hash
            - salt
          example:
            hash: U+Q43l2oMyLFXJBgZOGo2AF5d3sy0lWQYjqX+bVEwm/KwBWSP83OVRHlbi3tDc6KdOXc6N7JzOkK9s50PYOjEg==
            salt: 9qtiDMW9KczgmJUhYaj/n+IrS0LoWBNZFWARUZV1RQk=
        role:
          $ref: '#/components/schemas/Role'
        terminal:
          allOf:
            - $ref: '#/components/schemas/TerminalID'
            - nullable: true
        port:
          allOf:
            - $ref: '#/components/schemas/PortID'
            - nullable: true
      additionalProperties: false
      required:
        - email
        - key
        - role
        - terminal
        - port
      example:
        email: jdoe@tormails.com
        key:
          hash: U+Q43l2oMyLFXJBgZOGo2AF5d3sy0lWQYjqX+bVEwm/KwBWSP83OVRHlbi3tDc6KdOXc6N7JzOkK9s50PYOjEg==
          salt: 9qtiDMW9KczgmJUhYaj/n+IrS0LoWBNZFWARUZV1RQk=
        role: vessel planner
        terminal: 1
        port: null
    EmployeeList:
      title: Employee List
      description: An array of Employee objects.
      type: array
      items:
        $ref: '#/components/schemas/Employee'
      maxItems: 2147483647
      example:
        - email: jdoe@tormails.com
          key:
            hash: U+Q43l2oMyLFXJBgZOGo2AF5d3sy0lWQYjqX+bVEwm/KwBWSP83OVRHlbi3tDc6KdOXc6N7JzOkK9s50PYOjEg==
            salt: 9qtiDMW9KczgmJUhYaj/n+IrS0LoWBNZFWARUZV1RQk=
          role: vessel planner
          terminal: 1
          port: null
  responses:
    NoContent:
      description: 204 No Content
    NotFound:
      description: 404 Not Found
    Unauthorized:
      description: 401 Unauthorized
    Forbidden:
      description: 403 Forbidden
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: You must generate a JWT through the /token/obtain endpoint and then pass it as a Bearer token in the Authorization header.
...